agente_orchestrator/requirements.txt
google-adk>=0.4.0
litellm>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
tzdata>=2024.1
